chatbot-avatar/livekit_agent/requirements.txt
livekit-agents>=0.10.0
livekit-plugins-groq>=0.3.0
livekit-plugins-simli>=0.1.0
livekit-plugins-silero>=0.6.0
edge-tts>=6.1.9
miniaudio>=1.58
aiohttp>=3.9.0
python-dotenv>=1.0.0
